move: 도메인 테스트를 domain 디렉터리로 이동
