+++ src/test/java/domain/LottoGameTest.java
@@ -1,19 +1,15 @@
+package domain;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.within;
 
-import domain.LottoGame;
-import domain.LottoNumber;
-import domain.LottoPrice;
-import domain.LottoResult;
-import domain.LottoTicket;
-import domain.WinningLotto;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-@DisplayName("LottoResult 단위 테스트")
-public class LottoResultTest {
+@DisplayName("LottoGame 단위 테스트")
+class LottoGameTest {
 
     private final LottoTicket winningLottoTicket = new LottoTicket(List.of(1, 2, 3, 4, 5, 6));
     private final LottoNumber bonusNumber = new LottoNumber(7);
@@ -26,8 +22,8 @@ public class LottoResultTest {
             new LottoTicket(List.of(1, 2, 3, 4, 7, 8))
         );
 
-        LottoResult lottoResult = new LottoResult(lottoTickets, winningLotto);
-        Map<LottoPrice, Integer> result = lottoResult.getLottoResult();
+        LottoGame lottoGame = new LottoGame(winningLotto, lottoTickets);
+        Map<LottoPrice, Integer> result = lottoGame.getRank();
 
         assertThat(result).containsEntry(LottoPrice.NOTHING, 1);
         assertThat(result).containsEntry(LottoPrice.FOURTH, 1);
