+++ src/test/java/domain/LottoNumberTest.java
@@ -1,6 +1,7 @@
+package domain;
+
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
