+++ src/test/java/domain/LottoStoreTest.java
@@ -1,9 +1,8 @@
+package domain;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import domain.LottoStore;
-import domain.LottoTicket;
-import domain.Money;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
