+++ src/test/java/domain/MoneyTest.java
@@ -1,7 +1,7 @@
+package domain;
+
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import domain.Money;
-import domain.Quantity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
